pyproject.toml
[project]
name = "rag-recipes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "ipython>=8.37.0",
    "jinja2>=3.1.6",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-milvus>=0.3.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "matplotlib>=3.10.7",
    "milvus-lite>=2.5.1",
    "pymilvus>=2.6.4",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.1.2",
    "transformers>=4.57.3",
    "vllm>=0.11.2",
    "wandb>=0.23.0",
]
